Git commit 2 for issue number #18298 by aayush146

+++ ivy/functional/frontends/torch/creation_ops.py
@@ -299,12 +299,12 @@ def frombuffer(
 
 def empty_strided(*args,
                   stride,
-                  axis = None,
-                  size = None,
-                  dtype = None,
-                  device = None,
-                  requires_grad = False,
-                  memory_format = None):
+                  axis=None,
+                  size=None,
+                  dtype= None,
+                  device=None,
+                  requires_grad=False,
+                  memory_format=None):
 
     stride_list = list(stride)
     size_list = list(size)
@@ -312,5 +312,5 @@ def empty_strided(*args,
     del stride_list[axis] # removing an axis to avoid striding over
     del size_list[axis]
 
-    return ivy.empty_strided(agrs = args, stride = stride, axis = axis, size = size, dtype = dtype,
-                             device = device, requires_grad = requires_grad, memory_format = memory_format)
+    return ivy.empty_strided(agrs=args, stride=stride, axis=axis, size=size, dtype=dtype,
+                             device=device, requires_grad=requires_grad, memory_format=memory_format)
